NIO-866, addressing review comment by moving TestingModuleInitializer to its own file
